Use the data directory name to define the language in setup_new_wiki.py
Then now we should use "es" instead of "es_ES" for the data directory name.
This change do easier implement other wikis as simple_en or dialects.
In the case of need a es_es and another es_lat activities,
we will need rename the activity.info.lang name to be the same than
the directory used.

Signed-of-by: Gonzalo Odiard <gonzalo@laptop.org>