Use search index to validate existing links and check redirects - SL #3465 and OLPC #11837
We had a list of redirected pages, in a text file, and loaded it (twice!) in memory at activity start. This array was used to check if a article is offline or online, and to check redirects when display articles.

The load of this list was doing load the activity slowly, and in XOs with only 512MB crash the activity.

The redirect information was also charged in the database used to search,
then we had the same information 3 times.

This patch modify the data retriver to use the same database in the 3 cases. To do this, need add information to the database, then the old bundles need recreate the index removing the directory index_dir and running tools/create_index.py again.

This patch have a downside, searching in the index is slower than in a in memory array, then displaying the pages with many links will be slower than before. This will be adressed in the next patch.

Signed-off-by: Gonzalo Odiard <gonzalo@laptop.org>