Delay check links local/remote
With the last change, to use a database instead of a array to validate
redirections in links, the display of pages is a lot slower.
This patch implement the check after the page is displayed,
with a javascript code requesting a list of remote links using ajax,
and updating the url and class of this links in the page.

Signed-off-by: Gonzalo Odiard <gonzalo@laptop.org>