Use sqlite3 to search and obtain articles block and position
Previously used whoosh, but was not space efficient, and we need reduce
the space used by the activities. Using a sql database have additional
advantages, like checking existence of a list of links in the database
with only one query.
To use this change, we need recreate the index in the diferent languages.

Signed-off-by: Gonzalo Odiard <gonzalo@laptop.org>