Move initialization parameters for wikipedia activities
In the past, we created a module and class for every wikipedia version.
That class initilized configuration values, and imported from activity.py
This originated a lot of code duplication.
This patch moves the init parameters to a section in the activity.info file,
and activity.info read the config there. A module utils.py implement
the configuration read.
Also a test_server.py utility was added to run the server form the command line.
Can overrride the config using parameters, but by default will take the config
from activity.info file.

This change is implemented for the SimpleEnglish version,
to implement on other we need:
* Add a section 'Wikipedia' to the proper activity.info.<lang> file
and copy the parameters fron the activity_<lang>.py file.
* Modifiy exec line in the activity.info.<lang> file
to point to activity.WikipediaActivity
* Remove the file activity_<lang>.py