setup.py: if a data_file is not set in the command line take it from .info
We need do this to be able to use setup.py when packaging in debian.
In this way, setup.py is compatible with the standard use.
If you want change to create a different wikipedia language version
you can set the data file in the command line, and the activity.info file
will be overwriten.

Added messages to explaint to the user what is doing.